change clientdata to client and client to clientlist
